(PC-13076)[API] refactor: move import_dms_users from script to dms api
