+++ api/src/pcapi/core/subscription/exceptions.py
@@ -24,3 +24,10 @@ class BeneficiaryFraudCheckEligibilityTypeMismatchException(SubscriptionExceptio
 
 class CannotUpgradeBeneficiaryRole(SubscriptionException):
     pass
+
+
+class DMSParsingError(ValueError):
+    def __init__(self, user_email: str, errors: dict[str, str], *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.errors = errors
+        self.user_email = user_email
