+++ api/src/pcapi/scheduled_tasks/clock.py
@@ -19,6 +19,7 @@ from pcapi.core.offers.repository import delete_past_draft_offers
 from pcapi.core.offers.repository import find_tomorrow_event_stock_ids
 import pcapi.core.payments.utils as payments_utils
 from pcapi.core.providers.repository import get_provider_by_local_class
+from pcapi.core.subscription.dms import api as dms_api
 from pcapi.core.users import api as users_api
 from pcapi.core.users.external.user_automations import (
     users_beneficiary_credit_expiration_within_next_3_months_automation,
@@ -41,7 +42,6 @@ from pcapi.scheduled_tasks.decorators import cron_context
 from pcapi.scheduled_tasks.decorators import cron_require_feature
 from pcapi.scheduled_tasks.decorators import log_cron_with_transaction
 from pcapi.scripts.beneficiary import archive_dms_applications
-from pcapi.scripts.beneficiary import import_dms_users
 from pcapi.scripts.booking.handle_expired_bookings import handle_expired_bookings
 from pcapi.scripts.booking.notify_soon_to_be_expired_bookings import notify_soon_to_be_expired_individual_bookings
 from pcapi.scripts.payment.user_recredit import recredit_underage_users
@@ -80,7 +80,7 @@ def synchronize_provider_api() -> None:
 @log_cron_with_transaction
 def pc_import_dms_users_beneficiaries() -> None:
     procedure_id = settings.DMS_NEW_ENROLLMENT_PROCEDURE_ID
-    import_dms_users.run(procedure_id)
+    dms_api.import_dms_users(procedure_id)
     archive_dms_applications.archive_applications(procedure_id, dry_run=False)
 
 
@@ -92,7 +92,7 @@ def pc_import_dms_users_beneficiaries_from_old_dms() -> None:
     if not settings.IS_PROD:
         return
     procedure_id = 44623
-    import_dms_users.run(procedure_id)
+    dms_api.import_dms_users(procedure_id)
     archive_dms_applications.archive_applications(procedure_id, dry_run=False)
 
 
@@ -100,7 +100,7 @@ def pc_import_dms_users_beneficiaries_from_old_dms() -> None:
 @log_cron_with_transaction
 def pc_import_beneficiaries_from_dms_v3() -> None:
     procedure_id = settings.DMS_ENROLLMENT_PROCEDURE_ID_AFTER_GENERAL_OPENING
-    import_dms_users.run(procedure_id)
+    dms_api.import_dms_users(procedure_id)
     archive_dms_applications.archive_applications(procedure_id, dry_run=False)
 
 
@@ -114,7 +114,7 @@ def pc_import_beneficiaries_from_dms_v4() -> None:
         if not procedure_id:
             logger.info("Skipping DMS %s because procedure id is empty", procedure_name)
             continue
-        import_dms_users.run(procedure_id)
+        dms_api.import_dms_users(procedure_id)
         archive_dms_applications.archive_applications(procedure_id, dry_run=False)
